change int32 to long on ubuntu - 64bit related?

+++ model_c.pyx
@@ -4,8 +4,8 @@ import numpy.random as npr
 cimport cython
 
 @cython.boundscheck(False)
-def mutation(np.ndarray[np.int32_t, ndim=1, negative_indices=False] population, np.ndarray[double, ndim=1, negative_indices=False] mutation_rates):
-    cdef np.ndarray[np.int32_t, ndim=1, negative_indices=False] mutations
+def mutation(np.ndarray[long, ndim=1, negative_indices=False] population, np.ndarray[double, ndim=1, negative_indices=False] mutation_rates):
+    cdef np.ndarray[long, ndim=1, negative_indices=False] mutations
     mutations = npr.poisson(population*mutation_rates)
     cdef Py_ssize_t i
     for i in range(population.shape[0]-1):
